functions/update_order_status/requirements.txt
azure-functions
psycopg2-binary
requests
